update backEnd : createNode

+++ frontEnd/src/main/controller/modelController.ts
@@ -21,15 +21,20 @@ const upload = multer({ storage: storage })
 
 // api/model/register
 modelController.post('/register', upload.single('file'), async (req: Request, res: Response) => {
-  const fileDir = req.file!.path
+  const fileDir = req.file?.path
   const data = JSON.parse(req.body.data)
-  const result = await model_service.register(data, fileDir)
+  const parentId = req.body.parentId
+  const result = await model_service.createNode(data, parentId)
+  // TODO: 再调用具体的注册方法
+  // const result = await model_service.register(data, fileDir, parentId)
   // 删除上传的文件
-  fs.unlink(fileDir, (err) => {
-    if (err) {
-      console.error('Error deleting uploaded file:', err)
-    }
-  })
+  if (fileDir) {
+    fs.unlink(fileDir, (err) => {
+      if (err) {
+        console.error('Error deleting uploaded file:', err)
+      }
+    })
+  }
   res.status(result.code).json(result)
 })
 
@@ -39,14 +44,14 @@ modelController.get('/list', async (_req: Request, res: Response) => {
   res.status(result.code).json(result)
 })
 
-// api/model/delete/:id
-modelController.delete('/delete/:id', async (req: Request, res: Response) => {
+// api/model/:id
+modelController.delete('/:id', async (req: Request, res: Response) => {
   const result = await model_service.delete(req.params.id)
   res.status(result.code).json(result)
 })
 
-// api/model/update/:id
-modelController.put('/update/:id', async (req: Request, res: Response) => {
+// api/model/:id
+modelController.put('/:id', async (req: Request, res: Response) => {
   const result = await model_service.updateNode(req.params.id, req.body)
   res.status(result.code).json(result)
 })
